% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{preprocess_sample_subset}
\alias{preprocess_sample_subset}
\title{Preprocess weights for subset sampling.}
\usage{
preprocess_sample_subset(W)
}
\arguments{
\item{W}{Input weights where \code{W[i]} should be between 0 and 1.}
}
\value{
A datastructure that allows faster weighted random sampling through
        use of the preprocessing data structure.
}
\description{
Preprocess weights for subset sampling.
}
\examples{

# First create some data in (0,1]
W <- runif(100)

# Sample with preprocessing
structure <- preprocess_sample_subset(W)
idxs <- sample_subset(struct = structure)

}

