% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Perform Weighted Random Sampling}
\usage{
sample(A, W, k = 1)
}
\arguments{
\item{A}{A vector/list of indices to sample}

\item{W}{The weight of each corresponding index in A}

\item{k}{The number of indices to sample}
}
\value{
\code{k} indices sampled according to weights
}
\description{
Perform Weighted Random Sampling
}
\examples{
sample(1:100, rep(1/100, 100), k = 10)
}

