% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{sample_subset}
\alias{sample_subset}
\title{Perform subset sampling}
\usage{
sample_subset(W = NULL, struct = NULL, draws = NULL)
}
\arguments{
\item{W}{A vector of weights to sample from. Not needed if preprocessing has
been performed.}

\item{struct}{The preprocessed structured}

\item{draws}{Meant to specify the amount of draws you want, currently does
nothing.}
}
\value{
A vector of indices, where each index \code{i} is sampled with
        probability \code{W[i]}.
}
\description{
Perform subset sampling
}
\examples{

# First create some data in (0,1]
W <- runif(100)

# Sample indices from it
idxs <- sample_subset(W)

# .. or use preprocessing
structure <- preprocess_sample_subset(W)
idxs <- sample_subset(struct = structure)

}

