% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrs.R
\name{wrs_preprocess}
\alias{wrs_preprocess}
\title{Perform Weighted Random Sampling Preprocessing}
\usage{
wrs_preprocess(A, W, k = 1, method = "binary")
}
\arguments{
\item{A}{A vector/list of indices to sample}

\item{W}{The weight of each corresponding index in A}

\item{k}{The number of indices to sample}

\item{method}{Must be either 'binary' or 'rstree'. This determines the
sampling method used.}
}
\value{
\code{S} Prepared structure for sampling.
}
\description{
Perform Weighted Random Sampling Preprocessing
}

