% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{wrs_sample}
\alias{wrs_sample}
\title{Perform Weighted Random Sampling}
\usage{
wrs_sample(A = NULL, W = NULL, k = 1, method = "binary", struct)
}
\arguments{
\item{A}{A vector/list of indices to sample. This value is not needed if
struct has been initialized and passed.}

\item{W}{The weight of each corresponding index in A. This value is not
needed if struct has been initialized and passed.}

\item{k}{The number of indices to sample.}

\item{method}{Must be either 'binary' or 'rstree'. This determines the
sampling method used.}

\item{struct}{a preprocessed chunk of data.}
}
\value{
\code{k} indices sampled according to weights
}
\description{
Perform Weighted Random Sampling
}
\examples{
wrs_sample(1:100, rep(1/100, 100), k = 10, method='binary')

wrs_sample(1:100, rep(1/100, 100), k = 10, method='rstree')

}

