% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{wrs_sample}
\alias{wrs_sample}
\title{Perform Weighted Random Sampling}
\usage{
wrs_sample(A = NULL, W = NULL, k = 1, method = "binary", struct,
  draws = NULL)
}
\arguments{
\item{A}{A vector/list of indices to sample. This value is not needed if
struct has been initialized and passed.}

\item{W}{The weight of each corresponding index in A. This value is not
needed if struct has been initialized and passed.}

\item{k}{The number of indices to sample.}

\item{method}{Must be either 'binary' or 'rstree'. This determines the
sampling method used.}

\item{struct}{a preprocessed chunk of data.}

\item{draws}{The number of samples to draw. By default this method returns
a vector of indices. If draws is specified, the function
returns a matrix with dimensions (k, draws).}
}
\value{
\code{k} indices sampled according to weights
}
\description{
Perform Weighted Random Sampling
}
\examples{
wrs_sample(1:100, rep(1/100, 100), k = 10, method='binary')

wrs_sample(1:100, rep(1/100, 100), k = 10, method='rstree')

# To use preprocessing do the following: 
struct <- wrs_preprocess(1:100, rep(1/100, 100), k = 10)
samples <- wrs_sample(k=10, struct=struct)

# Do many draws in a single sampling operation, same as running a for loop
# and sampling 5 times, but faster. Returns a matrix.
struct <- wrs_preprocess(1:100, rep(1/100, 100), k = 10)
samples <- wrs_sample(k=10, struct=struct, draws=5)

}

